codex/refactor-server-modules (#55)
* Refactor Supervaizer server modules

* ✨ feat: update GitNexus index stats in AGENTS.md

* ✨docs update changelog AGENTS guidance

* ✨feat: add v2 methods, tests and license header
